vega/algorithms/nas/modnas/requirements.txt
numpy
pyyaml>=5.1
